Create new event for code folding
Create the new event `events.FOLDED` for manually manipulating the code folding
levels in the current buffer. This event is necessary because with current
facilities, it's not possible to add code folding support for some types of
separators which need to be folded, such as markdown frontmatter. I personally
implemented this feature.